+++ lib/pages/login_page.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import '../common/widgets/responsive_layout.dart';
+import '../mobile/login_page_mobile.dart';
+import '../pad/login_page_pad.dart';
+
+class LoginPage extends StatelessWidget {
+  const LoginPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const ResponsiveLayout(
+      mobile: LoginPageMobile(),
+      tablet: LoginPagePad(),
+    );
+  }
+}